feat: banner repository 분리
