+++ KnockKnock-iOS/Repository/HotPostRepository.swift
@@ -15,34 +15,9 @@ protocol HotPostRepositoryProtocol {
   func requestChallengeTitles(
     completionHandler: @escaping ([ChallengeTitle]) -> Void
   )
-  func requestBanner(bannerType: BannerType) async -> [HomeBanner]?
 }
 
 final class HotPostRepository: HotPostRepositoryProtocol {
-
-  /// 홈 메인/바 배너 조회
-  ///
-  /// - Parameters:
-  ///  - bannerType: 배너 형태(메인/바)
-  func requestBanner(bannerType: BannerType) async -> [HomeBanner]? {
-    do {
-
-      let result = try await KKNetworkManager
-        .shared
-        .asyncRequest(
-          object: ApiResponseDTO<[HomeBannerDTO]>.self,
-          router: .getBanner(bannerType: bannerType.rawValue)
-        )
-
-      guard let data = result.data else { return nil }
-      return data.map { $0.toDomain() }
-
-    } catch {
-      print(error)
-      
-      return nil
-    }
-  }
 
   /// 인기 게시글 조회
   func requestHotPost(
